tests(integration.WineDataParser): add new tests

+++ YggdrasilVinum.Tests/Unit/Parsers/TestHelper.cs
@@ -1,6 +1,3 @@
-using System;
-using System.Collections.Generic;
-using System.IO;
 using YggdrasilVinum.Models;
 
 namespace YggdrasilVinum.Tests.Unit.Parsers;
@@ -8,7 +5,7 @@ namespace YggdrasilVinum.Tests.Unit.Parsers;
 public static class TestHelper
 {
     public static readonly string ValidCsvHeader = "id,label,year,type";
-    
+
     public static readonly List<(int Id, string Label, int Year, WineType Type)> TestWineData = new()
     {
         (1, "Château Margaux", 2018, WineType.Red),
@@ -17,41 +14,39 @@ public static class TestHelper
         (4, "Biondi-Santi", 2015, WineType.Red),
         (5, "Domaine Tempier", 2021, WineType.Rose)
     };
-    
+
     public static string GenerateTestCsvContent(IEnumerable<(int Id, string Label, int Year, WineType Type)> entries)
     {
         using var stringWriter = new StringWriter();
         stringWriter.WriteLine(ValidCsvHeader);
-        
+
         foreach (var entry in entries)
         {
-            string typeStr = entry.Type switch
+            var typeStr = entry.Type switch
             {
                 WineType.Red => "tinto",
                 WineType.White => "branco",
                 WineType.Rose => "rosé",
                 _ => throw new ArgumentOutOfRangeException()
             };
-            
+
             stringWriter.WriteLine($"{entry.Id},{entry.Label},{entry.Year},{typeStr}");
         }
-        
+
         return stringWriter.ToString();
     }
-    
-    public static List<WineRecord> CreateWineRecords(IEnumerable<(int Id, string Label, int Year, WineType Type)> entries)
+
+    public static List<WineRecord> CreateWineRecords(
+        IEnumerable<(int Id, string Label, int Year, WineType Type)> entries)
     {
         var records = new List<WineRecord>();
-        foreach (var entry in entries)
-        {
-            records.Add(new WineRecord(entry.Id, entry.Label, entry.Year, entry.Type));
-        }
+        foreach (var entry in entries) records.Add(new WineRecord(entry.Id, entry.Label, entry.Year, entry.Type));
         return records;
     }
-    
+
     public static string GenerateInvalidTestCsvContent()
     {
-        return 
+        return
             $"{ValidCsvHeader}\n" +
             "1,Valid Wine,2020,tinto\n" +
             "not_an_id,Invalid ID,2021,tinto\n" +
@@ -61,19 +56,16 @@ public static class TestHelper
             "5,Too Few Fields,2023\n" +
             "6,Too Many Fields,2024,tinto,extra_field\n";
     }
-    
+
     public static string CreateTempCsvFile(string content)
     {
-        string tempFilePath = Path.GetTempFileName();
+        var tempFilePath = Path.GetTempFileName();
         File.WriteAllText(tempFilePath, content);
         return tempFilePath;
     }
-    
+
     public static void DeleteTempFile(string filePath)
     {
-        if (File.Exists(filePath))
-        {
-            File.Delete(filePath);
-        }
+        if (File.Exists(filePath)) File.Delete(filePath);
     }
 }
